Refactor sequence generation code
Use strings and regular expressions rather than lists and recursive programming to generate and filter the sequences of numbers that represent complex events.

Also switch from keras to tf.keras in an attempt to get the combined_loss function to work, but it doesn't.

Simplify the graphing that shows the distributions of complex event types.